CTM Accounts sync/requirements.txt
pandas>=1.5.3
pandas-gbq>=0.19.1
requests>=2.28.2
google-cloud-bigquery>=3.9.0
google-cloud-storage>=2.8.0
google-auth>=2.17.2
google-api-python-client>=2.166.0
python-dotenv>=1.0.0